requirements.txt
click==8.1.7
docutils==0.21.2
et-xmlfile==1.1.0
grpcio==1.63.0
grpcio-tools==1.63.0
numpy==1.26.4
openpyxl==3.1.2
overrides==7.7.0
pandas==2.2.2
protobuf==5.26.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
robotframework==7.0
robotframework-assertion-engine==3.0.3
robotframework-browser==18.4.0
robotframework-datadriver==1.11.1
robotframework-pythonlibcore==4.4.1
setuptools==69.5.1
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
wrapt==1.16.0
xlrd==2.0.1
